+++ app/src/main/java/com/sdex/activityrunner/about/LicensesDialogFragment.kt
@@ -7,10 +7,12 @@ import android.os.Message
 import android.webkit.WebChromeClient
 import android.webkit.WebView
 import android.webkit.WebViewClient
+import androidx.core.view.ViewCompat
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.sdex.activityrunner.R
 import com.sdex.activityrunner.commons.BaseDialogFragment
 import com.sdex.activityrunner.util.IntentUtils
+import com.sdex.activityrunner.about.LicenseThemeJSInterface
 
 class LicensesDialogFragment : BaseDialogFragment() {
 
@@ -45,13 +47,20 @@ class LicensesDialogFragment : BaseDialogFragment() {
                 onPositionChanged(t)
             }
         }
+        webView.addJavascriptInterface(LicenseThemeJSInterface(requireContext()), "theme")
         webView.webViewClient = object : WebViewClient() {
             override fun onPageFinished(view: WebView?, url: String?) {
                 super.onPageFinished(view, url)
                 view?.scrollTo(0, position)
             }
         }
+        ViewCompat.setScrollIndicators(
+            webView,
+            ViewCompat.SCROLL_INDICATOR_TOP or ViewCompat.SCROLL_INDICATOR_BOTTOM,
+            ViewCompat.SCROLL_INDICATOR_TOP or ViewCompat.SCROLL_INDICATOR_BOTTOM,
+        )
         webView.settings.setSupportMultipleWindows(true)
+        webView.settings.javaScriptEnabled = true
         webView.webChromeClient = object : WebChromeClient() {
             override fun onCreateWindow(
                 view: WebView, isDialog: Boolean,