+++ app/src/main/java/com/sdex/activityrunner/app/MainViewModel.kt
@@ -10,9 +10,11 @@ import com.sdex.activityrunner.db.cache.query.GetApplicationsQuery
 import com.sdex.activityrunner.preferences.AppPreferences
 import com.sdex.activityrunner.util.ApplicationsLoader
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
+import timber.log.Timber
 import javax.inject.Inject
 
 @HiltViewModel
@@ -26,8 +28,9 @@ class MainViewModel @Inject constructor(
     val searchQuery: LiveData<String?> = _searchQuery
 
     val items: LiveData<List<ApplicationModel>> = searchQuery.switchMap { text ->
-        val query = GetApplicationsQuery(appPreferences, text).sqLiteQuery
-        cacheRepository.getApplications(query)
+        val query = GetApplicationsQuery(appPreferences, text)
+        Timber.d("Query: $query")
+        cacheRepository.getApplications(query.sqLiteQuery)
     }
 
     init {
@@ -35,6 +38,7 @@ class MainViewModel @Inject constructor(
     }
 
     fun search(text: String?) {
+        Timber.d("Search: $text")
         _searchQuery.value = text
     }
 
@@ -42,6 +46,7 @@ class MainViewModel @Inject constructor(
         search(searchQuery.value)
     }
 
+    @OptIn(DelicateCoroutinesApi::class)
     private fun syncDatabase() {
         GlobalScope.launch(Dispatchers.IO) {
             applicationsLoader.syncDatabase()