+++ app/src/main/java/com/sdex/activityrunner/preferences/TooltipPreferences.kt
@@ -1,21 +1,19 @@
 package com.sdex.activityrunner.preferences
 
 import android.content.Context
-import android.content.SharedPreferences
-import kotlin.properties.Delegates
+import androidx.core.content.edit
 
 class TooltipPreferences(context: Context) {
 
-    private var sharedPreferences: SharedPreferences by Delegates.notNull()
-
-    init {
-        sharedPreferences = context.getSharedPreferences("tooltips_states", Context.MODE_PRIVATE)
-    }
+    private val sharedPreferences = context.getSharedPreferences(
+        "tooltips_states",
+        Context.MODE_PRIVATE,
+    )
 
     var showChangeIcon: Boolean
         get() = sharedPreferences.getBoolean(SHORTCUT_ICON_CHANGE, SHORTCUT_ICON_CHANGE_DEFAULT)
         set(show) {
-            sharedPreferences.edit().putBoolean(SHORTCUT_ICON_CHANGE, show).apply()
+            sharedPreferences.edit { putBoolean(SHORTCUT_ICON_CHANGE, show) }
         }
 
     companion object {