+++ app/src/main/java/com/sdex/activityrunner/shortcut/AddShortcutDialogActivity.kt
@@ -14,7 +14,6 @@ import androidx.activity.result.contract.ActivityResultContracts.PickVisualMedia
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.PopupMenu
 import androidx.core.graphics.drawable.toBitmap
-import androidx.core.view.isVisible
 import androidx.core.widget.doOnTextChanged
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
@@ -80,20 +79,22 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
                 .error(R.mipmap.ic_launcher)
                 .apply(RequestOptions().centerCrop())
                 .override(launcherLargeIconSize)
-                .into(object : CustomTarget<Drawable>() {
-                    override fun onResourceReady(
-                        resource: Drawable,
-                        transition: Transition<in Drawable>?,
-                    ) {
-                        bitmap = resource.toBitmap()
-                        binding.icon.setImageDrawable(resource)
-                        showTooltip()
-                    }
+                .into(
+                    object : CustomTarget<Drawable>() {
+                        override fun onResourceReady(
+                            resource: Drawable,
+                            transition: Transition<in Drawable>?,
+                        ) {
+                            bitmap = resource.toBitmap()
+                            binding.icon.setImageDrawable(resource)
+                            showTooltip()
+                        }
 
-                    override fun onLoadCleared(placeholder: Drawable?) {
-                    }
-                })
-            binding.useRoot.isVisible = true
+                        override fun onLoadCleared(placeholder: Drawable?) {
+                        }
+                    },
+                )
+            // TODO maybe check if root is available
             binding.useRoot.isChecked = !activityModel.exported
 
             binding.label.doOnTextChanged { _, _, _, count ->
@@ -106,7 +107,9 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
             binding.label.setText(activityModel.label)
             binding.label.text?.let { binding.label.setSelection(it.length) }
             binding.label.setSimpleItems(
-                setOf(activityModel.label, activityModel.name).filterNotNull().toTypedArray()
+                setOf(activityModel.label, activityModel.name)
+                    .filterNotNull()
+                    .toTypedArray(),
             )
         }
 
@@ -114,6 +117,14 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
             binding.label.setText(historyModel.name)
             binding.icon.setImageResource(R.mipmap.ic_launcher)
             binding.valueLayout.endIconMode = TextInputLayout.END_ICON_CLEAR_TEXT
+        }
+
+        binding.invertIconColors.setOnCheckedChangeListener { _, isChecked ->
+            binding.icon.colorFilter = if (isChecked) {
+                ShortcutColorInvertColorFilter()
+            } else {
+                null
+            }
         }
 
         binding.icon.setOnClickListener {
@@ -136,17 +147,19 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
             activityModel?.let {
                 val model = it.copy(name = shortcutName)
                 IntentUtils.createLauncherIcon(
-                    this,
-                    model,
-                    bitmap,
-                    binding.useRoot.isChecked
+                    context = this,
+                    activityModel = model,
+                    bitmap = bitmap,
+                    useRoot = binding.useRoot.isChecked,
+                    invertIconColors = binding.invertIconColors.isChecked,
                 )
             }
 
             historyModel?.let {
                 createHistoryModelShortcut(
-                    historyModel,
-                    shortcutName
+                    historyModel = historyModel,
+                    shortcutName = shortcutName,
+                    invertIconColors = binding.invertIconColors.isChecked,
                 )
             }
 
@@ -168,15 +181,16 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
                     binding.icon,
                     binding.content,
                     getString(R.string.shortcut_set_icon_tooltip),
-                    ToolTip.POSITION_BELOW
+                    ToolTip.POSITION_BELOW,
                 )
                 builder.setBackgroundColor(
                     resolveColorAttr(
-                        com.google.android.material.R.attr.colorTertiary
-                    )
+                        com.google.android.material.R.attr.colorTertiary,
+                    ),
                 )
                 builder.setTextAppearance(R.style.TooltipTextAppearance)
                 toolTipsManager.show(builder.build())
+
                 preferences.showChangeIcon = false
             }
         }
@@ -201,15 +215,17 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
                 pickMedia.launch(PickVisualMediaRequest(PickVisualMedia.ImageOnly))
             } else if (menuItem.itemId == R.id.pick_icon) {
                 if (iconPack != null) {
-                    val iconDialog = IconDialog.newInstance(IconDialogSettings {
-                        showSelectBtn = false
-                    })
+                    val iconDialog = IconDialog.newInstance(
+                        IconDialogSettings {
+                            showSelectBtn = false
+                        },
+                    )
                     iconDialog.show(supportFragmentManager, ICON_DIALOG_TAG)
                 } else {
                     Toast.makeText(
                         this,
                         R.string.icons_loading_error,
-                        Toast.LENGTH_SHORT
+                        Toast.LENGTH_SHORT,
                     ).show()
                 }
             }
@@ -218,12 +234,23 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
         popupMenu.show()
     }
 
-    private fun createHistoryModelShortcut(historyModel: HistoryModel, shortcutName: String) {
+    private fun createHistoryModelShortcut(
+        historyModel: HistoryModel,
+        shortcutName: String,
+        invertIconColors: Boolean = false,
+    ) {
         val historyToLaunchParamsConverter = HistoryToLaunchParamsConverter(historyModel)
         val launchParams = historyToLaunchParamsConverter.convert()
         val converter = LaunchParamsToIntentConverter(launchParams)
         val intent = converter.convert()
-        createShortcut(this, shortcutName, intent, bitmap)
+
+        createShortcut(
+            context = this,
+            name = shortcutName,
+            intent = intent,
+            icon = bitmap,
+            invertIconColors = invertIconColors,
+        )
     }
 
     private fun loadIcon(uri: Uri) {
@@ -234,15 +261,20 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
             .load(uri)
             .error(R.mipmap.ic_launcher)
             .apply(RequestOptions().centerCrop().override(size))
-            .into(object : CustomTarget<Bitmap>(size, size) {
-                override fun onResourceReady(resource: Bitmap, transition: Transition<in Bitmap>?) {
-                    bitmap = resource
-                    binding.icon.setImageBitmap(resource)
-                }
+            .into(
+                object : CustomTarget<Bitmap>(size, size) {
+                    override fun onResourceReady(
+                        resource: Bitmap,
+                        transition: Transition<in Bitmap>?,
+                    ) {
+                        bitmap = resource
+                        binding.icon.setImageBitmap(resource)
+                    }
 
-                override fun onLoadCleared(placeholder: Drawable?) {
-                }
-            })
+                    override fun onLoadCleared(placeholder: Drawable?) {
+                    }
+                },
+            )
     }
 
     companion object {
@@ -255,14 +287,16 @@ class AddShortcutDialogActivity : AppCompatActivity(), IconDialog.Callback {
             context.startActivity(
                 Intent(context, AddShortcutDialogActivity::class.java).apply {
                     putExtra(ARG_ACTIVITY_MODEL, activityModel)
-                }
+                },
             )
         }
 
         fun start(context: Context, historyModel: HistoryModel) {
-            context.startActivity(Intent(context, AddShortcutDialogActivity::class.java).apply {
-                putExtra(ARG_HISTORY_MODEL, historyModel)
-            })
+            context.startActivity(
+                Intent(context, AddShortcutDialogActivity::class.java).apply {
+                    putExtra(ARG_HISTORY_MODEL, historyModel)
+                },
+            )
         }
     }
 }