+++ app/src/main/java/com/sdex/activityrunner/tv/common/Dialog.kt
@@ -568,7 +568,7 @@ private fun Modifier.dialogFocusable() = composed {
     this.then(
         Modifier
             .focusRequester(focusRequester)
-            .focusProperties { exit = { FocusRequester.Cancel } }
+            .focusProperties { onExit = { FocusRequester.Cancel } }
             .focusGroup()
     )
 }